refactor: moved network package to the core module
