+++ app/src/main/java/com/monstarlab/features/login/domain/usecase/LoginUseCase.kt
@@ -1,6 +1,6 @@
 package com.monstarlab.features.login.domain.usecase
 
-import com.monstarlab.core.network.errorhandling.extensions.suspendRunCatching
+import com.monstarlab.core.extensions.suspendRunCatching
 import com.monstarlab.features.auth.domain.repository.AuthRepository
 import com.monstarlab.features.user.domain.repository.UserRepository
 import javax.inject.Inject
