+++ core/src/main/java/com/monstarlab/core/error/ErrorModelWithDisplayableMessage.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.network.errorhandling.error
+package com.monstarlab.core.error
 
 val ErrorModel.displayableMessage: String
     get() {
