+++ core/src/main/java/com/monstarlab/core/injection/CoroutinesModule.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.network.errorhandling.injection
+package com.monstarlab.core.injection
 
 import dagger.Module
 import dagger.Provides
