+++ core/src/main/java/com/monstarlab/core/network/errorhandling/ApiErrorDTO.kt
